+++ examples/M5Stack-SD-Menu/assets.h
@@ -761,7 +761,7 @@ const unsigned int ntp_jpeg_len = 835;
 
 
 
-unsigned char bluefork_jpg[]/* 10x14 */  = {
+const unsigned char bluefork_jpg[]/* 10x14 */  = {
   0xff, 0xd8, 0xff, 0xe0, 0x00, 0x10, 0x4a, 0x46, 0x49, 0x46, 0x00, 0x01,
   0x01, 0x01, 0x00, 0x60, 0x00, 0x60, 0x00, 0x00, 0xff, 0xdb, 0x00, 0x43,
   0x00, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
@@ -851,7 +851,7 @@ unsigned char bluefork_jpg[]/* 10x14 */  = {
 };
 unsigned int bluefork_jpg_len = 1022;
 
-unsigned char sd_updater32x40_jpg[] = {
+const unsigned char sd_updater32x40_jpg[] = {
   0xff, 0xd8, 0xff, 0xe0, 0x00, 0x10, 0x4a, 0x46, 0x49, 0x46, 0x00, 0x01,
   0x01, 0x01, 0x01, 0x2c, 0x01, 0x2c, 0x00, 0x00, 0xff, 0xdb, 0x00, 0x43,
   0x00, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
@@ -1009,7 +1009,7 @@ unsigned char sd_updater32x40_jpg[] = {
   0x89, 0x6b, 0x4e, 0x7f, 0xbc, 0xff, 0xd9
 };
 unsigned int sd_updater32x40_jpg_len = 1855;
-unsigned char sd_updater15x16_jpg[] = {
+const unsigned char sd_updater15x16_jpg[] = {
   0xff, 0xd8, 0xff, 0xe0, 0x00, 0x10, 0x4a, 0x46, 0x49, 0x46, 0x00, 0x01,
   0x01, 0x01, 0x00, 0x60, 0x00, 0x60, 0x00, 0x00, 0xff, 0xdb, 0x00, 0x43,
   0x00, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
